refactor: rename docker-plugins to docker-cli-plugins
