+++ docker-cli-plugins/commons/go.mod
@@ -1,0 +1,3 @@
+module github.com/Makepad-fr/wharf/docker-cli-plugins/commons
+
+go 1.22.2
