+++ docker-cli-plugins/commons/plugin_metadata.go
