+++ docker-cli-plugins/render/README.md
@@ -2,10 +2,10 @@
 
 ## Install the Docker plugin
 
-Use the following Makefile command to install the Docker plugin in your current Docker CLI installation
+Use the following command to install the Docker plugin in your current Docker CLI installation
 
 ```bash
-make install
+make install-plugins
 ```
 
 Test the installation using `docker render` from your Terminal.
