--- docker-plugins/README.md
@@ -1,6 +1,0 @@
-## Wharf Docker Plugin's
-
-Wharf can be installed as a set of Docker Plugin's and usable with Docker CLI using `docker <plugin_name>` syntax. To install the plugins run `make install-plugins` from the project's root.
-
-### Plugin List:
-- [docker-render](./render/README.md)
