--- docker-plugins/commons/go.mod
@@ -1,3 +1,0 @@
-module github.com/Makepad-fr/wharf/docker-plugins/commons
-
-go 1.22.2
