+++ go.work
@@ -3,6 +3,6 @@ go 1.22.2
 use (
 	./cli
 	./core
-	./docker-plugins/commons
-	./docker-plugins/render
+	./docker-cli-plugins/commons
+	./docker-cli-plugins/render
 )
